Fix: Bug when only one pin
When only one pin exists and has been clicked. On page reload, map zooms out to max
